Moved handlePullRequestPush to pull_request.go
